requirements.txt
# Core Data & Analysis
pandas>=1.5.0
numpy>=1.21.0
vectorbt>=0.25.0
scipy>=1.9.0

# Data Sources
yfinance>=0.2.0
ccxt>=4.0.0
alpha-vantage>=2.3.0

# Machine Learning & Optimization
scikit-learn>=1.1.0
optuna>=3.0.0

# Database & Storage
sqlalchemy>=1.4.0
h5py>=3.7.0

# Visualization & Reporting
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# Parallel Processing
joblib>=1.1.0

# Testing & Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
